oidc-postgres-demo/api-server/requirements.txt
Flask==3.0.0
requests==2.31.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
cryptography==41.0.7
